+++ src/CirculationLogList/CirculationLogList.js
@@ -1,4 +1,5 @@
 import React, {
+  useCallback,
   useMemo,
 } from 'react';
 import PropTypes from 'prop-types';
@@ -8,6 +9,9 @@ import {
 } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
 
+import {
+  useStripes,
+} from '@folio/stripes/core';
 import {
   Paneset,
   MultiColumnList,
@@ -25,6 +29,7 @@ import {
 import { useLocationFilters } from '../hooks';
 import { CirculationLogEventItems } from './CirculationLogEventItems';
 import { CirculationLogListFilter } from './CirculationLogListFilter';
+import { CirculationLogListActions } from './CirculationLogListActions';
 import { CirculationLogEventActions } from './CirculationLogEventActions';
 
 const resultsPaneTitle = <FormattedMessage id="ui-circulation-log.meta.title" />;
@@ -64,6 +69,7 @@ export const CirculationLogList = ({
   logEventsCount,
   servicePoints,
 }) => {
+  const stripes = useStripes();
   const history = useHistory();
   const location = useLocation();
 
@@ -84,6 +90,16 @@ export const CirculationLogList = ({
     }, {});
   }, [servicePoints]);
   const resultsFormatter = useMemo(() => getResultsFormatter(servicePointsMap), [servicePointsMap]);
+
+  const renderActionMenu = useCallback(
+    ({ onToggle }) => (
+      <CirculationLogListActions
+        logEventsCount={logEventsCount}
+        onToggle={onToggle}
+      />
+    ),
+    [logEventsCount],
+  );
 
   const resultsStatusMessage = (
     <NoResultsMessage
@@ -119,6 +135,7 @@ export const CirculationLogList = ({
         toggleFiltersPane={toggleFilters}
         filters={filters}
         isFiltersOpened={isFiltersOpened}
+        renderActionMenu={stripes.hasInterface('data-export-spring') && renderActionMenu}
       >
         <MultiColumnList
           id="circulation-log-list"